Deleted unused old SQL code. Migrated tests for DDL drop and create table operations to sqle package. Several are current failing.
